FotoVerifier/backend/api/cfa/src/requirements.txt
imageio
numpy
scipy
matplotlib
seaborn
Pillow
scikit-image
